+++ examples/ticketing/boot.ts
@@ -1,5 +1,5 @@
 import Fastify from 'fastify';
-import { startPlatform, setAccessToken } from '@unchainedshop/platform';
+import { startPlatform } from '@unchainedshop/platform';
 import baseModules from '@unchainedshop/plugins/presets/base.js';
 import connectBasePluginsToFastify from '@unchainedshop/plugins/presets/base-fastify.js';
 import { connect, unchainedLogger } from '@unchainedshop/api/lib/fastify/index.js';
@@ -36,7 +36,7 @@ try {
   });
 
   await seed(platform.unchainedAPI);
-  await setAccessToken(platform.unchainedAPI, 'admin', 'secret');
+  await platform.unchainedAPI.modules.users.setAccessToken('admin', 'secret');
 
   await fastify.listen({ host: '::', port: process.env.PORT ? parseInt(process.env.PORT) : 3000 });
 } catch (err) {