+++ packages/core/src/core-index.ts
@@ -27,6 +27,7 @@ export * from './bulk-importer/index.ts';
 export * from './services/index.ts';
 export * from './directors/index.ts';
 export * from './factory/index.ts';
+export { default as schedule, type ScheduleData } from './utils/schedule.ts';
 
 export interface UnchainedCoreOptions {
   db: mongodb.Db;